+++ .changeset/fast-papayas-heal.md
@@ -2,4 +2,4 @@
 "@pnpm/config": patch
 ---
 
-Symlink are not supported in the exFAT driver, and installation of dependencies will result in errors. Set the `node-linker` configuration in the exFAT driver to be `hoisted` by default.
+Symlink are not supported in the exFAT driver, and installation of dependencies will result in errors. The doctor command checks if the current drive is exFAT and gives a warning message.
